pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mtl-events"
version = "0.1.0"
description = "Montréal Events Agent - Curates and publishes events to Google Calendar"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.118.0",
    "beautifulsoup4>=4.12.0",
    "pytz>=2024.1",
    "pyyaml>=6.0.1",
    "click>=8.0.0",
    "feedparser>=6.0.0",
]

[project.scripts]
mtl-events = "src.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short" 